Add Unity desktop and normalize variable
This commit adds Unity as a possible Desktop when
getting the Icon Theme. Unity's settings are the
same as Gnome but in certain cases the current
desktop will distinguish between one or the other
one.

Additionally, the variable that controls which
desktop to look for themes for is normalized so
unity, Unity, UNITY are all considered the same
environment.